+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ### Breaking
 
+### Changes
+
+## [v0.8.4](https://github.com/malbeclabs/doublezero/compare/client/v0.8.3...client/v0.8.4) – 2026-01-28
+
+### Breaking
+
 - None for this release
 
 ### Changes
@@ -18,6 +24,7 @@ All notable changes to this project will be documented in this file.
   - Go SDK can now perform batch writes to device.health and link.health as per rfc12
 - Activator
   - fix(activator): add on-chain allocation support for users ([#2744](https://github.com/malbeclabs/doublezero/pull/2744))
+  - On-chain allocation enabled
 - Smartcontract
   - feat(smartcontract): add use_onchain_deallocation flag to MulticastGroup ([#2748](https://github.com/malbeclabs/doublezero/pull/2748))